Warning fix.
git-svn-id: http://svn.mythtv.org/svn/trunk@11308 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmyth/mythcontext.cpp
@@ -519,7 +519,7 @@ void MythContextPrivate::StoreGUIsettings()
     if (!font.exactMatch())
         font = QFont();
     font.setStyleHint(QFont::SansSerif, QFont::PreferAntialias);
-    font.setPointSize(floor(14 * m_hmult));
+    font.setPointSize((int)floor(14 * m_hmult));
 
     QApplication::setFont(font);
